requirements.txt
Django==2.2.19
djangorestframework==3.13.1
psycopg2-binary==2.8.6
python-dotenv==0.20.0
pytz==2022.1
sqlparse==0.4.2
